refactor: lift errors
